Align NewExtensionPage with CreateCodeMonitorPage
.. and the potentially new CreateBatchChangePage
