+++ client/web/src/enterprise/extensions/extension/RegistryExtensionForm.tsx
@@ -1,9 +1,9 @@
-import classNames from 'classnames'
 import * as React from 'react'
 
 import { ErrorAlert } from '@sourcegraph/branded/src/components/alerts'
 import { ErrorLike, isErrorLike } from '@sourcegraph/common'
 import { Scalars } from '@sourcegraph/shared/src/graphql-operations'
+import { Input, Select } from '@sourcegraph/wildcard'
 
 import {
     EXTENSION_NAME_MAX_LENGTH,
@@ -13,8 +13,6 @@ import {
 } from '../../../extensions/extension/extension'
 
 export const RegistryPublisherFormGroup: React.FunctionComponent<{
-    className?: string
-
     /** The current publisher value. */
     value?: Scalars['ID']
 
@@ -23,19 +21,19 @@ export const RegistryPublisherFormGroup: React.FunctionComponent<{
 
     disabled?: boolean
     onChange?: React.FormEventHandler<HTMLSelectElement>
-}> = ({ className = '', value, publishersOrError, disabled, onChange }) => (
-    <div className={classNames('form-group', className)}>
-        <label htmlFor="extension-registry-create-extension-page__publisher">Publisher</label>
+}> = ({ value, publishersOrError, disabled, onChange }) => (
+    <>
         {isErrorLike(publishersOrError) ? (
             <ErrorAlert error={publishersOrError} />
         ) : (
-            <select
-                id="extension-registry-create-extension-page__publisher"
-                className="form-control"
+            <Select
+                label="Publisher"
                 onChange={onChange}
                 required={true}
                 disabled={disabled || publishersOrError === 'loading'}
                 value={value}
+                aria-label="Publisher"
+                message="The owner of this extension. This can't be changed after creation."
             >
                 {publishersOrError === 'loading' ? (
                     <option disabled={true}>Loading...</option>
@@ -46,39 +44,31 @@ export const RegistryPublisherFormGroup: React.FunctionComponent<{
                         </option>
                     ))
                 )}
-            </select>
+            </Select>
         )}
-        <small className="form-help text-muted">
-            The owner of this extension. This can't be changed after creation.
-        </small>
-    </div>
+    </>
 )
 
 export const RegistryExtensionNameFormGroup: React.FunctionComponent<{
-    className?: string
     value: string
     disabled?: boolean
     onChange: React.FormEventHandler<HTMLInputElement>
-}> = ({ className = '', value, disabled, onChange }) => (
-    <div className={classNames('form-group', className)}>
-        <label htmlFor="extension-registry-form__name">Name</label>
-        <input
-            type="text"
-            name="extension-name"
-            className="form-control"
-            id="extension-registry-form__name"
-            onChange={onChange}
-            required={true}
-            autoFocus={true}
-            spellCheck={false}
-            autoCapitalize="off"
-            autoCorrect="off"
-            autoComplete="off"
-            value={value}
-            pattern={EXTENSION_NAME_VALID_PATTERN}
-            maxLength={EXTENSION_NAME_MAX_LENGTH}
-            disabled={disabled}
-        />
-        <small className="form-help text-muted">The name for this extension.</small>
-    </div>
+}> = ({ value, disabled, onChange }) => (
+    <Input
+        label="Name"
+        type="text"
+        name="extension-name"
+        onChange={onChange}
+        required={true}
+        autoFocus={true}
+        spellCheck={false}
+        autoCapitalize="off"
+        autoCorrect="off"
+        autoComplete="off"
+        value={value}
+        pattern={EXTENSION_NAME_VALID_PATTERN}
+        maxLength={EXTENSION_NAME_MAX_LENGTH}
+        disabled={disabled}
+        message="The name for this extension."
+    />
 )
